have traefik redirect http to https when https is available (#14)
This configures traefik to redirect any plain http traffic to https if Grist is configured to listen for that. Could be made
conditional on an environment variable, but it is hard to imagine a situation where this wouldn't be the desired behavior, so will wait for someone to ask for that.